+++ lib/llm/sequel/plugin.rb
@@ -320,19 +320,25 @@ module LLM::Sequel
 
     ##
     # @return [LLM::Provider]
+    # @raise [NotImplementedError]
+    #  when neither this model nor one of its ancestors implements the
+    #  callback
     def set_provider
+      return super if defined?(super)
       raise NotImplementedError, "implement the set_provider callback"
     end
 
     ##
     # @return [Hash]
     def set_context
+      return super if defined?(super)
       Plugin::EMPTY_HASH.dup
     end
 
     ##
     # @return [LLM::Tracer]
     def set_tracer
+      return super if defined?(super)
       nil
     end